p9: Refactor message piece types into their own file.
